Customized vagrant with ssh tunnels and host provisioning
The secret sauce is that Vagrantfiles are Ruby, so you can do plenty with it, including provisioning the host. In this case I create SSH tunnels so users don't have to remember the forwarded ports. Additionally, it updates the /etc/hosts file on the host for SSL and easy access in the browser.