setup webdav so eXist-db content is a virtual dir
Readme for project can be found at
https://github.com/grantmacken/nginx-eXist-ubuntu ( https://github.com/grantmacken/nginx-eXist-ubuntu )

This script sets up a webdav file system (davfs2 ) on your hardrive.

Once setup eXist can be mounted as a virtual file system. You can move around ,
copy and paste, drag and drop files into the database just as you would a normal
filesystem.

The file sysyem will be mounted by default in the users home dir in an
eXist dir. You can change this in the script with the line
davMount="/home/$SUDO_USER/eXist"

The script create a davfs group and adds the 'sudo user' to the group.
You have to login to make this effective.

To allow an ordinary user to mount there must be an entry in fstab so the script
does this. Altering fstab is dangerous business in linux based systems. We
make a copy of /etc/fstab as /etc/fstab_copy just in case.